Move and refactor scaffold logic
